user-service/requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
psycopg2-binary
bcrypt
flasgger
python-dotenv
requests
stripe
